New Problem
Created a new problem where the user has to find the numerical derivative of a simple polynomial function. Includes edge case when the exponent is zero.

Can add the number when it is assigned. Wanted to PR quickly to check whether this format for calculus questions is the best and adapt other stuff before moving onwards with other problems.